--- components/Common/ProgressionSidebar/ProgressionSidebarItem/index.module.css
@@ -1,40 +1,0 @@
-.item {
-  @apply relative
-    z-20
-    list-none
-    overflow-hidden
-    p-0;
-
-  a {
-    @apply flex
-      w-full
-      items-center
-      gap-1
-      text-sm
-      font-regular
-      text-neutral-800
-      hover:text-neutral-900
-      dark:text-neutral-200
-      dark:hover:text-white;
-  }
-
-  svg {
-    @apply flex-shrink-0
-      fill-neutral-200
-      stroke-white
-      stroke-[4]
-      dark:fill-neutral-800
-      dark:stroke-neutral-950;
-  }
-
-  &.active {
-    a {
-      @apply text-neutral-900
-        dark:text-white;
-    }
-
-    svg {
-      @apply fill-green-500;
-    }
-  }
-}
