+++ components/Containers/ProgressionSidebar/ProgressionSidebarGroup/index.module.css
@@ -1,25 +1,17 @@
-.group,
-ul {
-  @apply flex
-    list-none
-    flex-col
-    p-0;
-}
-
 .group {
-  & > li {
-    @apply flex
-      flex-col
-      gap-4
-      text-sm
-      font-medium
-      text-neutral-800
-      dark:text-neutral-200;
-  }
+  @apply flex
+    flex-col
+    gap-4
+    text-sm
+    font-medium
+    text-neutral-800
+    dark:text-neutral-200;
 
   .items {
     @apply relative
       -left-1
+      flex
+      flex-col
       gap-2;
 
     &::after {
@@ -34,7 +26,7 @@ ul {
         dark:bg-neutral-800;
     }
 
-    li {
+    a {
       &:first-child::before {
         @apply absolute
           bottom-[calc(50%+0.25rem)]
