+++ components/Containers/ProgressionSidebar/index.module.css
